Flow changes
- Debugging
- Removed hex data manipulation
- Added JSON data manipulation
- New program flow
- New game state detection and handling
- Added “interactive” mode where YOU are the Arduino

Python component boots and detects Arduino boot sequence and then
prompts for data from the Arduino. Data passing is not really working
yet, but the flow of the program is looking better and better. The
Python script now detects changes in the KSP game state and behaves
appropriately. Error handling makes everything a bit more graceful.