WIP: Arduino code
LockedInput is now working in a general sense. I want to put more logic
in there, but the button and switch work in conjunction and only three
pins are needed.

Issues:
Sometimes the lock doesn’t turn off correctly
Seems to be an issue reading the stage button during calibration

Also, removed an unnecessary read() to cut down on hardware polling
where not needed.