Joystick rework
Integrated button into joystick, pressing will cut sensitivity by half.
This bypasses the complication of enabling “soft controls.” The scale
has also been redone to be on a scale of -1 to 1 which is what the
Telemachus API expects. I also made it so when recalibrating, it
expands the range by 1 on either side. Also, added toString().