Fixed issues with avr-stl and compiled full sketch
Changed new.cpp to pnew.cpp and moved to the end of the first includes.
This fixed issues I was having with avr-stl.

Worked on Bargraph class to at least get it working enough to compile
into the main sketch. This involved a slight change to the Adafruit
backpack library to make it work. I have an issue open to see if that’s
something they will incorporate.

Uncommented all placeholder code and got it working up to compilation.
There be dragons.

Removed calibrate button since we’re not using it.
Remove ArduinoJson remnants from processInput() since we are manually
creating the JSON string

Remove const_ from iterators. Not needed.